imprv: User group detail page clean code (#5045)
* jsx -> tsx

* Typescriptized components & Wrap with unstated at the end to not use the return type of withTranslation in other files

* Typescriptized unstatedUtils

* Renamed state

* Explicit state type declaration

* Initial state as private member variable

* Added a null check

* Clean code

* Added comments for future tasks

* Fix lint

* Fix interface lint

* Replaced IUserGroup with IUserGroupHasObjectId

* Class Component to FC

* Class component to FC

* Removed unnecessary code

* Class component to FC

* Added a semicolon
